refactor gallery-stacked block
